添加 quartz可视化界面，cron 在线界面等

+++ modules/src/main/java/com/modules/quartz/dao/QuartzTaskInformationsMapper.java
@@ -17,7 +17,7 @@ public interface QuartzTaskInformationsMapper {
 
     QuartzTaskInformations selectByPrimaryKey(Long id);
 
-    int updateByPrimaryKeySelective(QuartzTaskInformations record);
+    Long updateByPrimaryKeySelective(QuartzTaskInformations record);
 
     int updateByPrimaryKey(QuartzTaskInformations record);
